chore(cleanup): Use index files and move module definitionss in folders
