--- nativescript-angular/directives.ts
@@ -1,42 +1,0 @@
-import { ListViewComponent, TemplateKeyDirective } from "./directives/list-view-comp";
-import { TabViewDirective, TabViewItemDirective } from "./directives/tab-view";
-import {
-    ActionBarComponent,
-    ActionBarScope,
-    ActionItemDirective,
-    NavigationButtonDirective
-} from "./directives/action-bar";
-import { AndroidFilterComponent, IosFilterComponent } from "./directives/platform-filters";
-
-export const NS_DIRECTIVES = [
-    ListViewComponent,
-    TemplateKeyDirective,
-    TabViewDirective,
-    TabViewItemDirective,
-    ActionBarComponent,
-    ActionBarScope,
-    ActionItemDirective,
-    NavigationButtonDirective,
-    AndroidFilterComponent,
-    IosFilterComponent,
-];
-
-export {
-    ListViewComponent,
-    SetupItemViewArgs,
-    TemplateKeyDirective
-} from "./directives/list-view-comp";
-export { TextValueAccessor } from "./value-accessors/text-value-accessor";
-export { CheckedValueAccessor } from "./value-accessors/checked-value-accessor";
-export { DateValueAccessor } from "./value-accessors/date-value-accessor";
-export { TimeValueAccessor } from "./value-accessors/time-value-accessor";
-export { NumberValueAccessor } from "./value-accessors/number-value-accessor";
-export { SelectedIndexValueAccessor } from "./value-accessors/selectedIndex-value-accessor";
-export { TabViewDirective, TabViewItemDirective } from "./directives/tab-view";
-export {
-    ActionBarComponent,
-    ActionBarScope,
-    ActionItemDirective,
-    NavigationButtonDirective
-} from "./directives/action-bar";
-export { AndroidFilterComponent, IosFilterComponent } from "./directives/platform-filters";
