+++ nativescript-angular/forms/forms.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
-import { TextValueAccessor } from "./value-accessors/text-value-accessor";
-import { CheckedValueAccessor } from "./value-accessors/checked-value-accessor";
-import { DateValueAccessor } from "./value-accessors/date-value-accessor";
-import { TimeValueAccessor } from "./value-accessors/time-value-accessor";
-import { NumberValueAccessor } from "./value-accessors/number-value-accessor";
-import { SelectedIndexValueAccessor } from "./value-accessors/selectedIndex-value-accessor";
+import {
+    TextValueAccessor,
+    CheckedValueAccessor,
+    DateValueAccessor,
+    TimeValueAccessor,
+    NumberValueAccessor,
+    SelectedIndexValueAccessor
+} from "./value-accessors";
 
 export const FORMS_DIRECTIVES = [
     TextValueAccessor,
