+++ nativescript-angular/forms/index.ts
@@ -1,0 +1,1 @@
+export * from "./forms.module";
